move to typescript, use pointer events, and update readme
