+++ apps/example-apps/react/examples/interaction/drag-and-drop/README.mdx
@@ -1,13 +1,14 @@
 ---
 title: Drag and Drop
-description: Drag and drop outside of the React Flow pane with native HTML Drag and Drop API or react-draggable.
+description: Drag and drop outside of the React Flow pane with native pointer events.
 ---
 
 A drag and drop user interface is very common for node-based workflow editors.
 The drag and drop behavior outside of the React Flow pane is not built in but
-can be implemented with the native [HTML Drag and Drop
-API](https://developer.mozilla.org/en-US/docs/Web/API/HTML_Drag_and_Drop_API)
-(used in this example) or a third party library like
-[react-draggable](https://github.com/react-grid-layout/react-draggable).
+can be implemented with the native [pointer events](https://developer.mozilla.org/en-US/docs/Web/API/Pointer_events).
+
+The reason why we use pointer events instead of the HTML Drag and Drop API is that
+the HTML Drag and Drop API does not provide support for touch devices.
+By using pointer events we can support both touch devices and the mouse.
 
 <RemoteCodeViewer route="examples/interaction/drag-and-drop" framework="react" />
