+++ apps/example-apps/react/examples/interaction/drag-and-drop/index.css
@@ -5,13 +5,13 @@ body {
   margin: 0;
   font-family: sans-serif;
   box-sizing: border-box;
-  /* Prevent text selection during drag on touch devices */
-  -webkit-touch-callout: none;
+  /* Prevent text selection during drag on touch devices
+  /* -webkit-touch-callout: none;
   -webkit-user-select: none;
   -khtml-user-select: none;
   -moz-user-select: none;
   -ms-user-select: none;
-  user-select: none;
+  user-select: none;  */
 }
 
 #app {
@@ -48,8 +48,8 @@ body {
   align-items: center;
   cursor: grab;
   /* Improve touch experience */
-  touch-action: none;
-  -webkit-tap-highlight-color: transparent;
+  /* touch-action: none; */
+  /* -webkit-tap-highlight-color: transparent; */
 }
 
 .dndflow .dndnode:active {
