+++ apps/example-apps/react/examples/interaction/drag-and-drop/index.tsx
@@ -4,6 +4,9 @@ import App from './App';
 import './index.css';
 
 const container = document.querySelector('#app');
+if (!container) {
+  throw new Error("Root container '#app' not found");
+}
 const root = createRoot(container);
 
 root.render(<App />);
